backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.7  # 添加这行
openai==1.3.5
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
python-dotenv==1.0.0
boto3==1.29.3
pydantic==2.5.2
pydantic-settings==2.1.0
reportlab==4.0.4 
requests==2.31.0